Split MessagePump into two parts
